Added portable_common directory. Common code for both PSA and PNA resides in it. Added common classes: PortableProgramStructure, ParsePortableArchitecture, InspectPortableProgram, PortableCodeGenerator, PortableOptions.
