+++ backends/bmv2/CMakeLists.txt
@@ -31,6 +31,17 @@ set (BMV2_SIMPLE_SWITCH_SRCS
     simple_switch/simpleSwitch.h
     simple_switch/options.h
     )
+    
+set (PORTABLE_COMMON
+    portable_common/portable.cpp
+    portable_common/portable.h  
+    portable_common/options.cpp
+    portable_common/options.h    
+    portable_common/midend.cpp
+    portable_common/midend.h
+    portable_common/portableProgramStructure.cpp
+    portable_common/portableProgramStructure.h
+    )
 
 set (BMV2_PSA_SWITCH_SRCS
     psa_switch/main.cpp
@@ -99,12 +110,15 @@ target_link_libraries (p4c-bm2-ss bmv2backend ${P4C_LIBRARIES} ${P4C_LIB_DEPS})
 
 install(TARGETS p4c-bm2-ss RUNTIME DESTINATION ${P4C_RUNTIME_OUTPUT_DIRECTORY})
 
+add_library(portable-common STATIC ${PORTABLE_COMMON})
+target_link_libraries(portable-common bmv2backend)
+
 add_executable(p4c-bm2-psa ${BMV2_PSA_SWITCH_SRCS})
-target_link_libraries (p4c-bm2-psa bmv2backend ${P4C_LIBRARIES} ${P4C_LIB_DEPS})
+target_link_libraries (p4c-bm2-psa portable-common ${P4C_LIBRARIES} ${P4C_LIB_DEPS})
 install(TARGETS p4c-bm2-psa RUNTIME DESTINATION ${P4C_RUNTIME_OUTPUT_DIRECTORY})
 
 add_executable(p4c-bm2-pna ${BMV2_PNA_NIC_SRCS})
-target_link_libraries (p4c-bm2-pna bmv2backend ${P4C_LIBRARIES} ${P4C_LIB_DEPS})
+target_link_libraries (p4c-bm2-pna portable-common ${P4C_LIBRARIES} ${P4C_LIB_DEPS})
 install(TARGETS p4c-bm2-pna RUNTIME DESTINATION ${P4C_RUNTIME_OUTPUT_DIRECTORY})
 
 file(RELATIVE_PATH
