+++ backends/bmv2/pna_nic/midend.cpp
@@ -90,7 +90,7 @@ class PnaEnumOn32Bits : public P4::ChooseEnumRepresentation {
     explicit PnaEnumOn32Bits(cstring filename) : filename(filename) {}
 };
 
-PnaNicMidEnd::PnaNicMidEnd(CompilerOptions &options, std::ostream *outStream) : MidEnd(options) {
+PnaNicMidEnd::PnaNicMidEnd(CompilerOptions &options, std::ostream *outStream) : PortableMidEnd(options) {
     auto convertEnums = new P4::ConvertEnums(&refMap, &typeMap, new PnaEnumOn32Bits("pna.p4"_cs));
     auto evaluator = new P4::EvaluatorPass(&refMap, &typeMap);
     std::function<bool(const Context *, const IR::Expression *)> policy =
@@ -122,12 +122,12 @@ PnaNicMidEnd::PnaNicMidEnd(CompilerOptions &options, std::ostream *outStream) : 
                 &refMap, &typeMap,
                 new P4::OrPolicy(new P4::IsValid(&refMap, &typeMap), new P4::IsMask())),
             new P4::ConstantFolding(&refMap, &typeMap),
-            new P4::StrengthReduction(&refMap, &typeMap),
+            new P4::StrengthReduction(&typeMap),
             new P4::SimplifySelectCases(&refMap, &typeMap, true),  // require constant keysets
             new P4::ExpandLookahead(&refMap, &typeMap),
             new P4::ExpandEmit(&refMap, &typeMap),
             new P4::SimplifyParsers(&refMap),
-            new P4::StrengthReduction(&refMap, &typeMap),
+            new P4::StrengthReduction(&typeMap),
             new P4::EliminateTuples(&refMap, &typeMap),
             new P4::SimplifyComparisons(&refMap, &typeMap),
             new P4::CopyStructures(&refMap, &typeMap),
@@ -142,12 +142,12 @@ PnaNicMidEnd::PnaNicMidEnd(CompilerOptions &options, std::ostream *outStream) : 
             new P4::ConstantFolding(&refMap, &typeMap),
             new P4::LocalCopyPropagation(&refMap, &typeMap, nullptr, policy),
             new PassRepeated({new P4::ConstantFolding(&refMap, &typeMap),
-                              new P4::StrengthReduction(&refMap, &typeMap)}),
+                              new P4::StrengthReduction(&typeMap)}),
             new P4::MoveDeclarations(),
             new P4::ValidateTableProperties({"pna_implementation"_cs, "pna_direct_counter"_cs,
                                              "pna_direct_meter"_cs, "pna_idle_timeout"_cs,
                                              "size"_cs}),
-            new P4::SimplifyControlFlow(&refMap, &typeMap),
+            new P4::SimplifyControlFlow(&typeMap),
             new P4::CompileTimeOperations(),
             new P4::TableHit(&refMap, &typeMap),
             new P4::EliminateSwitch(&refMap, &typeMap),
