+++ backends/bmv2/pna_nic/options.cpp
@@ -5,16 +5,4 @@
 
 namespace BMV2 {
 
-using namespace P4::literals;
-
-std::vector<const char *> *PnaNicOptions::process(int argc, char *const argv[]) {
-    searchForIncludePath(p4includePath,
-                         {"p4include/bmv2"_cs, "../p4include/bmv2"_cs, "../../p4include/bmv2"_cs},
-                         exename(argv[0]));
-
-    auto remainingOptions = CompilerOptions::process(argc, argv);
-
-    return remainingOptions;
-}
-
 }  // namespace BMV2
