+++ backends/bmv2/pna_nic/pnaProgramStructure.h
@@ -17,10 +17,7 @@ limitations under the License.
 #ifndef BACKENDS_BMV2_PNA_NIC_PNAPROGRAMSTRUCTURE_H_
 #define BACKENDS_BMV2_PNA_NIC_PNAPROGRAMSTRUCTURE_H_
 
-#include "backends/bmv2/common/backend.h"
-#include "backends/bmv2/common/programStructure.h"
-#include "ir/ir.h"
-#include "lib/cstring.h"
+#include "backends/bmv2/portable_common/portableProgramStructure.h"
 
 /// TODO: this is not really specific to BMV2, it should reside somewhere else
 namespace BMV2 {
@@ -31,59 +28,14 @@ enum pna_block_t {
     MAIN_DEPARSER,
 };
 
-class PnaProgramStructure : public ProgramStructure {
- protected:
-    P4::ReferenceMap *refMap;
-    P4::TypeMap *typeMap;
-
+class PnaProgramStructure : public PortableProgramStructure {
  public:
-    /// We place scalar user metadata fields (i.e., bit<>, bool)
-    /// in the scalars map.
-    ordered_map<cstring, const IR::Declaration_Variable *> scalars;
-    unsigned scalars_width = 0;
-    unsigned error_width = 32;
-    unsigned bool_width = 1;
-
     /// Architecture related information
     ordered_map<const IR::Node *, pna_block_t> block_type;
 
-    ordered_map<cstring, const IR::Type_Header *> header_types;
-    ordered_map<cstring, const IR::Type_Struct *> metadata_types;
-    ordered_map<cstring, const IR::Type_HeaderUnion *> header_union_types;
-    ordered_map<cstring, const IR::Declaration_Variable *> headers;
-    ordered_map<cstring, const IR::Declaration_Variable *> metadata;
-    ordered_map<cstring, const IR::Declaration_Variable *> header_stacks;
-    ordered_map<cstring, const IR::Declaration_Variable *> header_unions;
-    ordered_map<cstring, const IR::Type_Error *> errors;
-    ordered_map<cstring, const IR::Type_Enum *> enums;
-    ordered_map<cstring, const IR::P4Parser *> parsers;
-    ordered_map<cstring, const IR::P4ValueSet *> parse_vsets;
-    ordered_map<cstring, const IR::P4Control *> deparsers;
-    ordered_map<cstring, const IR::P4Control *> pipelines;
-    ordered_map<cstring, const IR::Declaration_Instance *> extern_instances;
-    ordered_map<cstring, cstring> field_aliases;
-
-    std::vector<const IR::ExternBlock *> globals;
-
  public:
     PnaProgramStructure(P4::ReferenceMap *refMap, P4::TypeMap *typeMap)
-        : refMap(refMap), typeMap(typeMap) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
-    }
-
-    std::set<cstring> non_pipeline_controls;
-    std::set<cstring> pipeline_controls;
-
-    bool hasVisited(const IR::Type_StructLike *st) {
-        if (auto h = st->to<IR::Type_Header>())
-            return header_types.count(h->getName());
-        else if (auto s = st->to<IR::Type_Struct>())
-            return metadata_types.count(s->getName());
-        else if (auto u = st->to<IR::Type_HeaderUnion>())
-            return header_union_types.count(u->getName());
-        return false;
-    }
+        : PortableProgramStructure(refMap, typeMap) {}
 
     /// Checks if a string is of type PNA_CounterType_t returns true
     /// if it is, false otherwise.
@@ -98,7 +50,7 @@ class PnaProgramStructure : public ProgramStructure {
     }
 };
 
-class ParsePnaArchitecture : public Inspector {
+class ParsePnaArchitecture : public ParsePortableArchitecture {
     PnaProgramStructure *structure;
 
  public:
@@ -112,7 +64,6 @@ class ParsePnaArchitecture : public Inspector {
                 node->getNode());
     }
 
-    bool preorder(const IR::ToplevelBlock *block) override;
     bool preorder(const IR::PackageBlock *block) override;
     bool preorder(const IR::ExternBlock *block) override;
 
@@ -123,16 +74,13 @@ class ParsePnaArchitecture : public Inspector {
     }
 };
 
-class InspectPnaProgram : public Inspector {
-    P4::ReferenceMap *refMap;
-    P4::TypeMap *typeMap;
+class InspectPnaProgram : public InspectPortableProgram {
     PnaProgramStructure *pinfo;
 
  public:
     InspectPnaProgram(P4::ReferenceMap *refMap, P4::TypeMap *typeMap, PnaProgramStructure *pinfo)
-        : refMap(refMap), typeMap(typeMap), pinfo(pinfo) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
+    : InspectPortableProgram(refMap, typeMap),
+      pinfo(pinfo) {
         CHECK_NULL(pinfo);
         setName("InspectPnaProgram");
     }
@@ -141,7 +89,6 @@ class InspectPnaProgram : public Inspector {
     void postorder(const IR::P4Control *c) override;
     void postorder(const IR::Declaration_Instance *di) override;
 
-    bool isHeaders(const IR::Type_StructLike *st);
     void addTypesAndInstances(const IR::Type_StructLike *type, bool meta);
     void addHeaderType(const IR::Type_StructLike *st);
     void addHeaderInstance(const IR::Type_StructLike *st, cstring name);
