+++ backends/bmv2/portable_common/midend.cpp
@@ -1,0 +1,69 @@
+/*
+Copyright 2024 Marvell Technology, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+#include "midend.h"
+
+#include "backends/bmv2/common/check_unsupported.h"
+#include "backends/bmv2/portable_common/options.h"
+#include "frontends/common/constantFolding.h"
+#include "frontends/common/resolveReferences/resolveReferences.h"
+#include "frontends/p4/evaluator/evaluator.h"
+#include "frontends/p4/fromv1.0/v1model.h"
+#include "frontends/p4/moveDeclarations.h"
+#include "frontends/p4/simplify.h"
+#include "frontends/p4/simplifyParsers.h"
+#include "frontends/p4/strengthReduction.h"
+#include "frontends/p4/typeChecking/typeChecker.h"
+#include "frontends/p4/typeMap.h"
+#include "frontends/p4/uniqueNames.h"
+#include "frontends/p4/unusedDeclarations.h"
+#include "midend/actionSynthesis.h"
+#include "midend/compileTimeOps.h"
+#include "midend/complexComparison.h"
+#include "midend/convertEnums.h"
+#include "midend/copyStructures.h"
+#include "midend/eliminateInvalidHeaders.h"
+#include "midend/eliminateNewtype.h"
+#include "midend/eliminateSerEnums.h"
+#include "midend/eliminateSwitch.h"
+#include "midend/eliminateTuples.h"
+#include "midend/expandEmit.h"
+#include "midend/expandLookahead.h"
+#include "midend/fillEnumMap.h"
+#include "midend/flattenHeaders.h"
+#include "midend/flattenInterfaceStructs.h"
+#include "midend/local_copyprop.h"
+#include "midend/midEndLast.h"
+#include "midend/nestedStructs.h"
+#include "midend/orderArguments.h"
+#include "midend/predication.h"
+#include "midend/removeAssertAssume.h"
+#include "midend/removeLeftSlices.h"
+#include "midend/removeMiss.h"
+#include "midend/removeSelectBooleans.h"
+#include "midend/removeUnusedParameters.h"
+#include "midend/replaceSelectRange.h"
+#include "midend/simplifyKey.h"
+#include "midend/simplifySelectCases.h"
+#include "midend/simplifySelectList.h"
+#include "midend/tableHit.h"
+#include "midend/validateProperties.h"
+
+namespace BMV2 {
+
+using namespace P4::literals;
+
+}  // namespace BMV2
