+++ backends/common/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(BACKENDS_COMMON_SRCS
     metermap.cpp
     programStructure.cpp
+    ../bmv2/portable_common/portableProgramStructure.cpp
     psaProgramStructure.cpp
 )
