+++ backends/common/psaProgramStructure.cpp
@@ -31,16 +31,6 @@ void InspectPsaProgram::postorder(const IR::Declaration_Instance *di) {
     }
 }
 
-bool InspectPsaProgram::isHeaders(const IR::Type_StructLike *st) {
-    bool result = false;
-    for (auto f : st->fields) {
-        if (f->type->is<IR::Type_Header>() || f->type->is<IR::Type_Stack>()) {
-            result = true;
-        }
-    }
-    return result;
-}
-
 void InspectPsaProgram::addHeaderType(const IR::Type_StructLike *st) {
     LOG5("In addHeaderType with struct " << st->toString());
     if (st->is<IR::Type_HeaderUnion>()) {
@@ -224,14 +214,6 @@ void InspectPsaProgram::postorder(const IR::P4Control *c) {
         else if (info.first == EGRESS && info.second == DEPARSER)
             pinfo->deparsers.emplace("egress"_cs, c);
     }
-}
-
-bool ParsePsaArchitecture::preorder(const IR::ToplevelBlock *block) {
-    // Blocks are not in IR tree, use a custom visitor to traverse.
-    for (auto it : block->constantValue) {
-        if (it.second->is<IR::Block>()) visit(it.second->getNode());
-    }
-    return false;
 }
 
 bool ParsePsaArchitecture::preorder(const IR::ExternBlock *block) {
