+++ backends/common/psaProgramStructure.h
@@ -18,11 +18,7 @@ limitations under the License.
 #ifndef BACKENDS_COMMON_PSAPROGRAMSTRUCTURE_H_
 #define BACKENDS_COMMON_PSAPROGRAMSTRUCTURE_H_
 
-#include "backends/common/programStructure.h"
-#include "frontends/common/resolveReferences/referenceMap.h"
-#include "frontends/p4/typeMap.h"
-#include "ir/ir.h"
-#include "lib/cstring.h"
+#include "backends/bmv2/portable_common/portableProgramStructure.h"
 
 namespace P4 {
 
@@ -33,59 +29,14 @@ enum block_t {
     DEPARSER,
 };
 
-class PsaProgramStructure : public ProgramStructure {
- protected:
-    P4::ReferenceMap *refMap;
-    P4::TypeMap *typeMap;
-
+class PsaProgramStructure : public BMV2::PortableProgramStructure {
  public:
-    /// We place scalar user metadata fields (i.e., bit<>, bool)
-    /// in the scalars map.
-    ordered_map<cstring, const IR::Declaration_Variable *> scalars;
-    unsigned scalars_width = 0;
-    unsigned error_width = 32;
-    unsigned bool_width = 1;
-
     /// Architecture related information.
     ordered_map<const IR::Node *, std::pair<gress_t, block_t>> block_type;
 
-    ordered_map<cstring, const IR::Type_Header *> header_types;
-    ordered_map<cstring, const IR::Type_Struct *> metadata_types;
-    ordered_map<cstring, const IR::Type_HeaderUnion *> header_union_types;
-    ordered_map<cstring, const IR::Declaration_Variable *> headers;
-    ordered_map<cstring, const IR::Declaration_Variable *> metadata;
-    ordered_map<cstring, const IR::Declaration_Variable *> header_stacks;
-    ordered_map<cstring, const IR::Declaration_Variable *> header_unions;
-    ordered_map<cstring, const IR::Type_Error *> errors;
-    ordered_map<cstring, const IR::Type_Enum *> enums;
-    ordered_map<cstring, const IR::P4Parser *> parsers;
-    ordered_map<cstring, const IR::P4ValueSet *> parse_vsets;
-    ordered_map<cstring, const IR::P4Control *> deparsers;
-    ordered_map<cstring, const IR::P4Control *> pipelines;
-    ordered_map<cstring, const IR::Declaration_Instance *> extern_instances;
-    ordered_map<cstring, cstring> field_aliases;
-
-    std::vector<const IR::ExternBlock *> globals;
-
  public:
     PsaProgramStructure(P4::ReferenceMap *refMap, P4::TypeMap *typeMap)
-        : refMap(refMap), typeMap(typeMap) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
-    }
-
-    std::set<cstring> non_pipeline_controls;
-    std::set<cstring> pipeline_controls;
-
-    bool hasVisited(const IR::Type_StructLike *st) {
-        if (auto h = st->to<IR::Type_Header>())
-            return header_types.count(h->getName());
-        else if (auto s = st->to<IR::Type_Struct>())
-            return metadata_types.count(s->getName());
-        else if (auto u = st->to<IR::Type_HeaderUnion>())
-            return header_union_types.count(u->getName());
-        return false;
-    }
+        : PortableProgramStructure(refMap, typeMap) {}
 
     /// Checks if a string is of type PSA_CounterType_t returns true
     /// if it is, false otherwise.
@@ -104,7 +55,7 @@ class PsaProgramStructure : public ProgramStructure {
     }
 };
 
-class ParsePsaArchitecture : public Inspector {
+class ParsePsaArchitecture : public BMV2::ParsePortableArchitecture {
     PsaProgramStructure *structure;
 
  public:
@@ -118,7 +69,6 @@ class ParsePsaArchitecture : public Inspector {
                 node->getNode());
     }
 
-    bool preorder(const IR::ToplevelBlock *block) override;
     bool preorder(const IR::PackageBlock *block) override;
     bool preorder(const IR::ExternBlock *block) override;
 
@@ -129,16 +79,13 @@ class ParsePsaArchitecture : public Inspector {
     }
 };
 
-class InspectPsaProgram : public Inspector {
-    P4::ReferenceMap *refMap;
-    P4::TypeMap *typeMap;
+class InspectPsaProgram : public BMV2::InspectPortableProgram {
     PsaProgramStructure *pinfo;
 
  public:
     InspectPsaProgram(P4::ReferenceMap *refMap, P4::TypeMap *typeMap, PsaProgramStructure *pinfo)
-        : refMap(refMap), typeMap(typeMap), pinfo(pinfo) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
+    : InspectPortableProgram(refMap, typeMap),
+      pinfo(pinfo) {
         CHECK_NULL(pinfo);
         setName("InspectPsaProgram");
     }
@@ -147,7 +94,6 @@ class InspectPsaProgram : public Inspector {
     void postorder(const IR::P4Control *c) override;
     void postorder(const IR::Declaration_Instance *di) override;
 
-    bool isHeaders(const IR::Type_StructLike *st);
     void addTypesAndInstances(const IR::Type_StructLike *type, bool meta);
     void addHeaderType(const IR::Type_StructLike *st);
     void addHeaderInstance(const IR::Type_StructLike *st, cstring name);
